logout functionality fixed
