+++ src/app/auth/join/join.component.spec.ts
@@ -8,7 +8,7 @@ describe('JoinComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ JoinComponent ]
+      declarations: [ JoinComponent ],
     })
     .compileComponents();
   }));
