+++ src/app/auth/logout/logout.component.html
@@ -1,3 +1,10 @@
-<p>
-  logout works!
-</p>
+<div class="landing-container-root">
+    <div class="center-block">
+        <div class="title">
+            <h1>GOOD FOOT CLUB</h1>
+        </div>
+        <div class="data-container">
+            <h3>You've been logged out</h3>
+        </div>
+    </div>
+</div>
