+++ src/app/auth/logout/logout.component.spec.ts
@@ -8,7 +8,7 @@ describe('LogoutComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ LogoutComponent ]
+      declarations: [ LogoutComponent ],
     })
     .compileComponents();
   }));
