+++ src/app/auth/logout/logout.component.ts
@@ -1,15 +1,23 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { Observable } from 'rxjs/Rx';
+import {WindowRefService} from '../../common/services/window.service';
 
 @Component({
   selector: 'app-logout',
   templateUrl: './logout.component.html',
-  styleUrls: ['./logout.component.styl']
+  styleUrls: ['./logout.component.styl'],
 })
 export class LogoutComponent implements OnInit {
 
-  constructor() { }
-
-  ngOnInit() {
+  constructor(
+      private windowRef: WindowRefService,
+      private router: Router,
+  ) {
+      this.windowRef.setFullScreen(true);
   }
 
+  ngOnInit() {
+      Observable.timer(3000).subscribe(() => this.router.navigate(['/']));
+  }
 }
