+++ src/app/auth/signin/signin.component.html
@@ -49,6 +49,10 @@
                     </div>
                 </button>
             </div>
+
+            <div class="back-button-container">
+                <a (click)="back()" class="back-button">back</a>
+            </div>
         </div>
     </div>
 </div>
