+++ src/app/auth/signin/signin.component.spec.ts
@@ -8,7 +8,7 @@ describe('SigninComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ SigninComponent ]
+      declarations: [ SigninComponent ],
     })
     .compileComponents();
   }));
