--- src/app/home/landing-page/login/login.component.html
@@ -1,47 +1,0 @@
-<div class="login-container">
-    <div class="signup-header">Please Sign In (Up)</div>
-
-    <form [formGroup]="loginForm" autocomplete="off" class="white-form">
-        <md-input-container #usernameContainer>
-            <input mdInput
-                   #username
-                   [style.width]="'100%'"
-                   autocomplete="off"
-                   formControlName="username"
-                   placeholder="Username"
-                   type='text'>
-            <md-hint align="end" *ngIf="usernameContainer._mdInputChild.focused">
-                <span *ngIf="!loginForm.controls.username.errors?.required">{{ username.value.length }} / 30</span>
-            </md-hint>
-        </md-input-container>
-
-        <md-input-container #passwordContainer>
-            <input mdInput
-                   #password
-                   [style.width]="'100%'"
-                   autocomplete="off"
-                   formControlName="password"
-                   placeholder="Password"
-                   type='password'>
-            <md-hint align="end" *ngIf="passwordContainer._mdInputChild.focused">
-                <span *ngIf="!loginForm.controls.password.errors?.required">{{ password.value.length }} / 30</span>
-            </md-hint>
-        </md-input-container>
-    </form>
-    <div class="button-container">
-        <button class="gffc-button btn-dark" (click)="loginUsingCredentials()">Sign In with Credentials</button>
-    </div>
-
-    <div class="button-container">
-        <button class="gffc-button btn-dark" (click)="useFacebook()">
-            <div class="icon-wrapper">
-                <svg class="facebook-icon" viewBox="0 0 95 95" preserveAspectRatio>
-                    <path d="M75.2-0.3L62.2-0.4c-14.6 0-24 9.7-24 24.7v11.4H25.1c-1.1 0-2 0.9-2 2v16.5c0 1.1 0.9 2 2 2h13.1v41.6c0 1.1 0.9 2 2 2h17c1.1 0 2-0.9 2-2V56.2h15.3c1.1 0 2-0.9 2-2l0-16.5c0-0.5-0.2-1.1-0.6-1.4 -0.4-0.4-0.9-0.6-1.4-0.6H59.3v-9.6c0-4.6 1.1-7 7.1-7l8.8 0c1.1 0 2-0.9 2-2V1.7C77.2 0.6 76.3-0.3 75.2-0.3z"/>
-                </svg>
-            </div>
-            <div class="text-wrapper">
-                <span>Sign In with Facebook</span>
-            </div>
-        </button>
-    </div>
-</div>
