--- src/app/home/landing-page/login/login.component.styl
@@ -1,19 +1,0 @@
-@import "../../../common/styles/variables.styl"
-@import "../../../common/styles/custom-buttons.styl"
-
-.login-container
-    display flex
-    flex-flow column
-    align-items center
-    justify-content center
-    min-height 200px
-
-    .signup-header
-        color $general-background-color
-        font-size $header-font-size
-        text-transform uppercase
-        text-align center
-        padding-bottom 30px
-
-    form
-        padding 0px 20px
