--- src/app/home/landing-page/register/register.component.styl
@@ -1,15 +1,0 @@
-@import "../../../common/styles/variables.styl"
-@import "../../../common/styles/custom-buttons.styl"
-
-.register-container
-    display flex
-    flex-flow column
-    align-items center
-    justify-content center
-    min-height 200px
-
-    form
-        padding 0px 20px
-
-.hint-container > span
-    padding-left 20px
